+++ test/foundry/utils/helpers/StakingTestHelpers.t.sol
@@ -79,9 +79,7 @@ contract StakingTestHelpers is StakingV2Setup {
                             V1 Helper Functions
     //////////////////////////////////////////////////////////////*/
 
-    function createRewardEscrowEntryV1(address _account, uint256 _amount)
-        internal
-    {
+    function createRewardEscrowEntryV1(address _account, uint256 _amount) internal {
         vm.prank(treasury);
         kwenta.approve(address(rewardEscrowV1), _amount);
         vm.prank(treasury);
@@ -197,7 +195,7 @@ contract StakingTestHelpers is StakingV2Setup {
     }
 
     function stakeEscrowedFundsV2(address _account, uint256 _amount) internal {
-        if (_amount != 0) createRewardEscrowEntryV2(_account, _amount, 52 weeks);
+        if (_amount != 0) createRewardEscrowEntryV2(_account, uint128(_amount), 52 weeks);
         vm.prank(_account);
         stakingRewardsV2.stakeEscrow(_amount);
     }
@@ -212,13 +210,11 @@ contract StakingTestHelpers is StakingV2Setup {
         stakingRewardsV2.unstakeEscrowSkipCooldown(_account, _amount);
     }
 
-    function createRewardEscrowEntryV2(address _account, uint256 _amount)
-        internal
-    {
+    function createRewardEscrowEntryV2(address _account, uint256 _amount) internal {
         vm.prank(treasury);
         kwenta.approve(address(rewardEscrowV2), _amount);
         vm.prank(treasury);
-        rewardEscrowV2.createEscrowEntry(_account, _amount, 52 weeks, 90);
+        rewardEscrowV2.createEscrowEntry(_account, uint128(_amount), 52 weeks, 90);
     }
 
     function createRewardEscrowEntryV2(address _account, uint256 _amount, uint256 _duration)
@@ -227,7 +223,7 @@ contract StakingTestHelpers is StakingV2Setup {
         vm.prank(treasury);
         kwenta.approve(address(rewardEscrowV2), _amount);
         vm.prank(treasury);
-        rewardEscrowV2.createEscrowEntry(_account, _amount, _duration, 90);
+        rewardEscrowV2.createEscrowEntry(_account, uint128(_amount), uint32(_duration), 90);
     }
 
     function createRewardEscrowEntryV2(
@@ -239,14 +235,16 @@ contract StakingTestHelpers is StakingV2Setup {
         vm.prank(treasury);
         kwenta.approve(address(rewardEscrowV2), _amount);
         vm.prank(treasury);
-        rewardEscrowV2.createEscrowEntry(_account, _amount, _duration, _earlyVestingFee);
+        rewardEscrowV2.createEscrowEntry(
+            _account, uint128(_amount), uint32(_duration), _earlyVestingFee
+        );
     }
 
     function appendRewardEscrowEntryV2(address _account, uint256 _amount) internal {
         vm.prank(treasury);
         kwenta.transfer(address(rewardEscrowV2), _amount);
         vm.prank(address(stakingRewardsV2));
-        rewardEscrowV2.appendVestingEntry(_account, _amount);
+        rewardEscrowV2.appendVestingEntry(_account, uint128(_amount));
     }
 
     function getStakingRewardsV2(address _account) internal {
